+++ tests/test_base_dfs.py
@@ -2,9 +2,10 @@ import pandas as pd
 import pytest
 
 # noinspection PyProtectedMember
+from typeddfs.base_dfs import UnsupportedOperationError
 from typeddfs.untyped_dfs import UntypedDf
 
-from . import TypedTrivial, sample_data, sample_data_2, sample_data_str
+from . import Trivial, sample_data, sample_data_2, sample_data_str
 
 
 class TestBase:
@@ -44,27 +45,27 @@ class TestBase:
             df.only("none", exclude_na=True)
 
     def test_cfirst(self):
-        df = TypedTrivial(sample_data())
+        df = Trivial(sample_data())
         assert df.column_names() == ["abc", "123", "xyz"]
         df2 = df.cfirst(["xyz", "123", "abc"])
         assert df2.column_names() == ["xyz", "123", "abc"]
         assert df.column_names() == ["abc", "123", "xyz"]
         assert df.cfirst("xyz").column_names() == ["xyz", "abc", "123"]
         # empty DF
-        assert TypedTrivial(df[df["abc"] == 9999]).cfirst("123").column_names() == [
+        assert Trivial(df[df["abc"] == 9999]).cfirst("123").column_names() == [
             "123",
             "abc",
             "xyz",
         ]
 
     def test_sort_col(self):
-        df = TypedTrivial.convert(pd.DataFrame(sample_data_str()))
+        df = Trivial.convert(pd.DataFrame(sample_data_str()))
         df2 = df.sort_natural("abc")
         assert df2.index_names() == []
         assert df2.index.tolist() == [1, 0]  # reversed
 
     def test_drop_cols(self):
-        df = TypedTrivial(sample_data())
+        df = Trivial(sample_data())
         df2 = df.drop_cols(["abc", "123"])
         assert list(df.columns) == ["abc", "123", "xyz"]
         assert list(df2.columns) == ["xyz"]
@@ -72,32 +73,32 @@ class TestBase:
         assert list(df3.columns) == ["abc", "123", "xyz"]
 
     def test_no_detype(self):
-        df = TypedTrivial(sample_data())
-        assert isinstance(df, TypedTrivial)
-        assert isinstance(df.reset_index(), TypedTrivial)
-        assert isinstance(df.reindex(), TypedTrivial)
-        assert isinstance(df.sort_natural("abc"), TypedTrivial)
-        assert isinstance(df.sort_values("abc"), TypedTrivial)
-        assert isinstance(df.copy(), TypedTrivial)
-        assert isinstance(df.abs(), TypedTrivial)
-        assert isinstance(df.drop_duplicates(), TypedTrivial)
-        assert isinstance(df.bfill(0), TypedTrivial)
-        assert isinstance(df.ffill(0), TypedTrivial)
-        assert isinstance(df.replace(123, 1), TypedTrivial)
-        assert isinstance(df.applymap(lambda s: s), TypedTrivial)
-        assert isinstance(df.drop(1), TypedTrivial)
-        assert isinstance(df.astype(str), TypedTrivial)
-        assert isinstance(df.drop("abc", axis=1), TypedTrivial)
-        assert isinstance(df.dropna(), TypedTrivial)
-        assert isinstance(df.fillna(0), TypedTrivial)
-        assert isinstance(df.append(df), TypedTrivial)
-        assert isinstance(df.rename(columns=dict(abc="twotwentytwo")), TypedTrivial)
+        df = Trivial(sample_data())
+        assert isinstance(df, Trivial)
+        assert isinstance(df.reset_index(), Trivial)
+        assert isinstance(df.reindex(), Trivial)
+        assert isinstance(df.sort_natural("abc"), Trivial)
+        assert isinstance(df.sort_values("abc"), Trivial)
+        assert isinstance(df.copy(), Trivial)
+        assert isinstance(df.abs(), Trivial)
+        assert isinstance(df.drop_duplicates(), Trivial)
+        assert isinstance(df.bfill(0), Trivial)
+        assert isinstance(df.ffill(0), Trivial)
+        assert isinstance(df.replace(123, 1), Trivial)
+        assert isinstance(df.applymap(lambda s: s), Trivial)
+        assert isinstance(df.drop(1), Trivial)
+        assert isinstance(df.astype(str), Trivial)
+        assert isinstance(df.drop("abc", axis=1), Trivial)
+        assert isinstance(df.dropna(), Trivial)
+        assert isinstance(df.fillna(0), Trivial)
+        assert isinstance(df.append(df), Trivial)
+        assert isinstance(df.rename(columns=dict(abc="twotwentytwo")), Trivial)
 
     def test_set_index(self):
         df = UntypedDf().convert(pd.DataFrame(sample_data()).set_index("abc"))
         assert df.set_index([]).index_names() == []
         assert df.set_index([], append=True).index_names() == ["abc"]
-        with pytest.warns(UserWarning):
+        with pytest.raises(UnsupportedOperationError):
             df.set_index([], inplace=True)