+++ tests/test_example.py
@@ -7,11 +7,12 @@ from typeddfs.example import run
 
 
 class TestExample:
-    cap = io.StringIO()
-    with contextlib.redirect_stdout(cap):
-        run()
-    cap = cap.getvalue().strip()
-    assert cap == "['key'] ['value', 'note']\n123"
+    def test(self):
+        cap = io.StringIO()
+        with contextlib.redirect_stdout(cap):
+            run()
+        cap = cap.getvalue().strip()
+        assert cap == "['key'] ['value', 'note']\n123"
 
 
 if __name__ == "__main__":