+++ tests/test_typed.py
@@ -6,10 +6,10 @@ from typeddfs.base_dfs import AbsDf, AsymmetricDfError
 from typeddfs.untyped_dfs import UntypedDf
 
 from . import (
-    TypedMultiIndex,
-    TypedSingleIndex,
+    Ind2,
+    Ind1,
     TypedSymmetric,
-    TypedTrivial,
+    Trivial,
     sample_data,
     sample_data_str,
     sample_symmetric_df,
@@ -23,47 +23,47 @@ class TestTyped:
         )
 
     def test_vanilla(self):
-        df = TypedTrivial.convert(pd.DataFrame(sample_data()))
+        df = Trivial.convert(pd.DataFrame(sample_data()))
         df2 = df.vanilla()
-        assert isinstance(df, TypedTrivial)
+        assert isinstance(df, Trivial)
         assert isinstance(df2, pd.DataFrame)
         assert not isinstance(df2, AbsDf)
 
     def test_convert_fail(self):
         with pytest.raises(TypeError):
-            TypedTrivial.convert(55)
+            Trivial.convert(55)
 
     def test_detype(self):
-        df = TypedTrivial.convert(pd.DataFrame(sample_data()))
+        df = Trivial.convert(pd.DataFrame(sample_data()))
         df2 = df.untyped()
-        assert isinstance(df, TypedTrivial)
+        assert isinstance(df, Trivial)
         assert isinstance(df2, UntypedDf)
 
     def test_is_multindex(self):
-        assert not TypedTrivial.convert(pd.DataFrame(sample_data())).is_multindex()
-        assert not TypedSingleIndex.convert(pd.DataFrame(sample_data())).is_multindex()
-        assert TypedMultiIndex.convert(pd.DataFrame(sample_data())).is_multindex()
+        assert not Trivial.convert(pd.DataFrame(sample_data())).is_multindex()
+        assert not Ind1.convert(pd.DataFrame(sample_data())).is_multindex()
+        assert Ind2.convert(pd.DataFrame(sample_data())).is_multindex()
 
     def test_lengths(self):
-        df = TypedSingleIndex.convert(pd.DataFrame(sample_data()))
+        df = Ind1.convert(pd.DataFrame(sample_data()))
         assert df.n_columns() == 2
         assert df.n_indices() == 1
         assert df.n_rows() == 2
 
     def test_sort_no_index(self):
-        df = TypedTrivial.convert(pd.DataFrame(sample_data_str()))
+        df = Trivial.convert(pd.DataFrame(sample_data_str()))
         df2 = df.sort_natural_index()
         assert df2.index_names() == []
 
     def test_sort_single_index(self):
-        df = TypedSingleIndex.convert(pd.DataFrame(sample_data_str()))
+        df = Ind1.convert(pd.DataFrame(sample_data_str()))
         df2 = df.sort_natural_index()
         assert df2.column_names() == ["123", "xyz"]
         assert df2.index_names() == ["abc"]
         assert df2.index.tolist() == ["bbb", "zzz"]
 
     def test_sort_multiindex(self):
-        df = TypedMultiIndex.convert(pd.DataFrame(sample_data_str()))
+        df = Ind2.convert(pd.DataFrame(sample_data_str()))
         df2 = df.sort_natural_index()
         assert df2.column_names() == ["123"]
         assert df2.index_names() == ["abc", "xyz"]
@@ -73,34 +73,34 @@ class TestTyped:
         # should be in place
         df = pd.DataFrame(sample_data())
         assert df.__class__.__name__ == "DataFrame"
-        TypedTrivial._change(df)
-        assert df.__class__.__name__ == "TypedTrivial"
+        Trivial._change(df)
+        assert df.__class__.__name__ == "Trivial"
 
     def test_not_inplace(self):
         df = pd.DataFrame(sample_data())
-        df2 = TypedMultiIndex(df)
-        df3 = TypedMultiIndex.convert(df)
+        df2 = Ind2(df)
+        df3 = Ind2.convert(df)
         assert df.__class__.__name__ == "DataFrame"
-        assert df2.__class__.__name__ == "TypedMultiIndex"
-        assert df3.__class__.__name__ == "TypedMultiIndex"
+        assert df2.__class__.__name__ == "Ind2"
+        assert df3.__class__.__name__ == "Ind2"
 
     def test_index_names(self):
-        df = TypedMultiIndex.convert(pd.DataFrame(sample_data()))
+        df = Ind2.convert(pd.DataFrame(sample_data()))
         assert df.index_names() == ["abc", "xyz"]
-        df = TypedTrivial.convert(pd.DataFrame(sample_data()))
+        df = Trivial.convert(pd.DataFrame(sample_data()))
         assert isinstance(df.index_names(), list)
         assert df.index_names() == []
 
     def test_column_names(self):
-        df = TypedTrivial(sample_data())
+        df = Trivial(sample_data())
         # df.columns == [...] would fail because it would resolve to array==array, which is ambiguous
         assert isinstance(df.column_names(), list)
         assert df.column_names() == ["abc", "123", "xyz"]
 
     def test_isvalid(self):
         df = pd.DataFrame(sample_data())
-        assert TypedMultiIndex.is_valid(df)
-        assert not TypedMultiIndex.is_valid(df.drop("abc", axis=1))
+        assert Ind2.is_valid(df)
+        assert not Ind2.is_valid(df.drop("abc", axis=1))
 
     def test_symmetric(self):
         df = sample_symmetric_df()